+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Change Log
 Not released
 ------------
 
-- Added support for [GPAW](https://gpaw.readthedocs.io/). All-electron wave function reconstruction is enabled.
+- Added support for [GPAW](https://gpaw.readthedocs.io/). All-electron wavefunction reconstruction is enabled.
 
 v2.3 (2025/12/20)
 -----------------